pyproject.toml
[project]
name = "parser"
version = "0.1.0"
description = ""
authors = [
    {name = "NKTKLN",email = "nktkln@nktkln.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "bs4 (>=0.0.2,<0.0.3)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
interrogate = "^1.7.0"
vulture = "^2.14"
pip-audit = "^2.9.0"
black = "^25.1.0"

[tool.black]
line-length = 88
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
  | tmp
)/
'''

[tool.ruff]
line-length = 88
target-version = "py39"
exclude = [".git", "__pycache__", "build", "dist", ".venv", "tmp"]

[tool.ruff.lint]
ignore = ["E203", "D100"]
select = ["E", "F", "W", "C90", "B", "B9", "I", "D", "S"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
disallow_untyped_defs = true
disallow_any_unimported = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
strict_optional = true
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
pretty = true
show_error_codes = true
exclude = [".git", "__pycache__", "build", "dist", ".venv", "tmp"]

[tool.interrogate]
fail-under = 90
exclude = [".git", "__pycache__", "build", "dist", ".venv", "tmp"]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
verbose = 1
color = true

[tool.vulture]
exclude = [".git", "__pycache__", "build", "dist", ".venv", "tmp"]
